lib-shared: split services into interface/implementation
