+++ Lib-Shared/src/test/java/com/kdgregory/example/javalambda/services/metadata/TestPhotoMetadata.java
@@ -13,9 +13,9 @@ import com.amazonaws.services.dynamodbv2.document.Item;
 
 import static net.sf.kdgcommons.test.NumericAsserts.*;
 
-import com.kdgregory.example.javalambda.shared.services.metadata.Fields;
-import com.kdgregory.example.javalambda.shared.services.metadata.PhotoMetadata;
-import com.kdgregory.example.javalambda.shared.services.metadata.Sizes;
+import com.kdgregory.example.javalambda.shared.data.Fields;
+import com.kdgregory.example.javalambda.shared.data.PhotoMetadata;
+import com.kdgregory.example.javalambda.shared.data.Sizes;
 
 import org.junit.Test;
 import static org.junit.Assert.*;
